requirements.txt
# Web framework
fastapi==0.95.1
uvicorn==0.23.2
pydantic==1.10.7
python-dotenv==1.0.0
httpx==0.24.1
python-multipart==0.0.6

# Database
sqlalchemy==2.0.15
alembic==1.11.1

# Utilities
loguru==0.7.0
uuid==1.30
statistics==1.0.3.5

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.1.0